+++ source/extensions/isaacsim.app.about/docs/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## [2.0.5] - 2024-12-07
+### Fixed
+- moved from deprecated editor_menu.add_item to omni.kit.menu.utils
+
 ## [2.0.4] - 2024-11-19
 ### Fixed
 - Startup test