+++ source/extensions/isaacsim.code_editor.jupyter/docs/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## [1.0.4] - 2024-12-07
+### Fixed
+- moved from deprecated editor_menu.add_item to omni.kit.menu.utils
+
 ## [1.0.3] - 2024-11-19
 ### Fixed
 - Startup test